Modified logging to unify what used to be three different logging sources: install-progress.yml, current.yml, and uploaded logs to the server. Instead of all this nonsense, we are now unifying all important status data into status.yml (but still keeping basic log information in bombardier.log). The current.yml file gets uploaded to the server rather routinely, keeping the server up to speed on what's going on. I removed that BOM file that gets saved on the client so that the client pulls it down directly from the server when it needs it. This could result in a performance penalty on the client. The BOM file format has now adopted YAML, as it was due to.
We are handling "file not found" errors differently from "server offline" errors, which was long overdue. 

Judging by the fact that integration testing used to take 90 seconds and now takes about 450, there is some performance tuning yet to be done. I'm thinking the server class needs to cache files that are downloaded and we need to be a bit more discriminating when we upload status to the server.